25991 - Correção do bug

+++ src/SME.SGP.Api/Controllers/PlanejamentoAnualController.cs
@@ -56,7 +56,7 @@ namespace SME.SGP.Api.Controllers
         [HttpGet("turmas/copia")]
         [ProducesResponseType(typeof(IEnumerable<TurmaParaCopiaPlanoAnualDto>), 200)]
         [ProducesResponseType(typeof(RetornoBaseDto), 500)]
-        public async Task<IActionResult> ObterTurmasParaCopia([FromQuery] long turmaId, [FromQuery] long componenteCurricular, [FromQuery] bool ensinoEspecial, [FromServices] IObterTurmasParaCopiaUseCase useCase)
+        public async Task<IActionResult> ObterTurmasParaCopia([FromQuery] int turmaId, [FromQuery] long componenteCurricular, [FromQuery] bool ensinoEspecial, [FromServices] IObterTurmasParaCopiaUseCase useCase)
         {
             return Ok(await useCase.Executar(turmaId, componenteCurricular, ensinoEspecial ));
         }
